removed redundant macros, moved structured output tests into a separate file
