+++ src/cpp/src/sampling/sampler.cpp
@@ -939,11 +939,10 @@ SamplerOutput Sampler::sample(const std::vector<SequenceGroup::Ptr> & sequence_g
         }
         #endif
         if (!m_logit_processors.count(request_id)) {
-            m_logit_processors.insert({request_id, LogitProcessor(sampling_params, sequence_group->get_prompt_ids()
-            #ifdef ENABLE_XGRAMMAR
-                                                                  , std::make_shared<StructuredOutputController>(m_tokenizer, m_tokenizer.get_vocab_vector().size())
-            #endif
-                                                                )});
+            if (!m_structured_output_controller) {
+                m_structured_output_controller = std::make_shared<StructuredOutputController>(m_tokenizer, m_tokenizer.get_vocab_vector().size());
+            }
+            m_logit_processors.insert({request_id, LogitProcessor(sampling_params, sequence_group->get_prompt_ids(), m_structured_output_controller)});
         }
         if (!m_stop_strings.count(request_id)) {
             auto processed_stop_string = process_stop_strings(sampling_params.stop_strings, m_tokenizer);
@@ -1017,21 +1016,12 @@ LogitProcessor& Sampler::get_logit_processor(uint64_t request_id) {
 
 void Sampler::create_logit_processor(uint64_t request_id, const GenerationConfig& sampling_params, const TokenIds& prompt) {
     std::cerr << "Creating logit processor for request id: " << request_id << std::endl;
-    #ifdef ENABLE_XGRAMMAR
-    if (sampling_params.is_structured_output_generation() && !m_structured_output_controller) {
+    if (!m_structured_output_controller) {
         std::cerr << "Structured output controller is created" << std::endl;
         m_structured_output_controller = std::make_shared<StructuredOutputController>(m_tokenizer, m_tokenizer.get_vocab_vector().size());
     }
-    #endif
 
-    m_logit_processors.insert(
-        {request_id,
-        LogitProcessor(sampling_params,
-                       prompt
-                       #ifdef ENABLE_XGRAMMAR
-                       ,m_structured_output_controller
-                        #endif
-                       )});
+    m_logit_processors.insert({request_id, LogitProcessor(sampling_params, prompt, m_structured_output_controller)});
 }
 
 void Sampler::clear_request_info(uint64_t request_id) {
