+++ tests/python_tests/data/models.py
@@ -8,8 +8,7 @@ from os.path import sep
 
 def get_models_list():
     precommit_models = [
-        "TinyLlama/TinyLlama-1.1B-Chat-v1.0",
-        # "katuni4ka/tiny-random-phi3",
+        "katuni4ka/tiny-random-phi3",
     ]
 
     nightly_models = [
